+++ test/deployment/light/deploymentStandaloneLight.test.js
@@ -48,8 +48,8 @@ describe('CMTAT Core - Standalone', function () {
       .withArgs(this.address1, ZERO_ADDRESS, VALUE1)
     // Emits a Burn event
     await expect(this.logs)
-      .to.emit(this.cmtat, 'Enforcement')
-      .withArgs(sender, this.address1, VALUE1, REASON_EVENT)
+      .to.emit(this.cmtat, 'ForcedTransfer(address,address,address,uint256,bytes)')
+      .withArgs(sender, this.address1, ZERO_ADDRESS, VALUE1, REASON_EVENT)
     // Check balances and total supply
     expect(await this.cmtat.balanceOf(this.address1)).to.equal(DIFFERENCE)
     expect(await this.cmtat.totalSupply()).to.equal(DIFFERENCE)
@@ -67,8 +67,8 @@ describe('CMTAT Core - Standalone', function () {
       .withArgs(this.address1, ZERO_ADDRESS, DIFFERENCE)
     // Emits a Burn event
     await expect(this.logs)
-      .to.emit(this.cmtat, 'Enforcement')
-      .withArgs(this.admin, this.address1, DIFFERENCE, REASON_EVENT)
+      .to.emit(this.cmtat, 'ForcedTransfer(address,address,address,uint256,bytes)')
+      .withArgs(this.admin, this.address1, ZERO_ADDRESS, DIFFERENCE, REASON_EVENT)
     // Check balances and total supply
     expect(await this.cmtat.balanceOf(this.address1)).to.equal(0)
     expect(await this.cmtat.totalSupply()).to.equal(0)